requirements.txt
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.0.0
pillow>=10.0.0
chromadb>=0.4.24
pypdf>=4.0.0
python-magic>=0.4.27
tqdm>=4.66.0
click>=8.1.7